Revert the try from tag u8 macro back to a simple TryFrom<u8> implementation since error types are consolidated
