+++ instruction-macros/crates/instruction-macros-impl/src/render/mod.rs
@@ -5,11 +5,11 @@ mod feature_namespace;
 mod instruction_accounts;
 mod instruction_data;
 mod pack_into_slice_trait;
-mod try_from_tag_macro;
+mod try_from_u8;
 
 pub use feature::*;
 pub use feature_namespace::*;
 pub use instruction_accounts::render as render_instruction_accounts;
 pub use instruction_data::render as render_instruction_data;
 pub use pack_into_slice_trait::render as render_pack_into_slice_trait;
-pub use try_from_tag_macro::render as render_try_from_tag_macro;
+pub use try_from_u8::render as render_try_from_u8;
