+++ instruction-macros/crates/instruction-macros/src/derive/instruction_data.rs
@@ -9,14 +9,14 @@ use instruction_macros_impl::{
     render::{
         render_instruction_data,
         render_pack_into_slice_trait,
-        render_try_from_tag_macro,
+        render_try_from_u8,
     },
 };
 use proc_macro2::TokenStream;
 use syn::DeriveInput;
 
 pub struct DeriveInstructionData {
-    pub try_from_u8_macro: TokenStream,
+    pub try_from_u8: TokenStream,
     pub pack_into_slice_trait: TokenStream,
     pub instruction_data: TokenStream,
 }
@@ -25,12 +25,12 @@ pub fn derive_instruction_data(input: DeriveInput) -> syn::Result<DeriveInstruct
     let parsed_enum = ParsedEnum::try_from((false, input))?;
     let instruction_variants = parse_instruction_variants(&parsed_enum)?;
 
-    let try_from_u8_macro = render_try_from_tag_macro(&parsed_enum, &instruction_variants);
+    let try_from_u8 = render_try_from_u8(&parsed_enum, &instruction_variants);
     let instruction_data = render_instruction_data(&parsed_enum, instruction_variants);
     let pack_into_slice_trait = render_pack_into_slice_trait();
 
     Ok(DeriveInstructionData {
-        try_from_u8_macro,
+        try_from_u8,
         pack_into_slice_trait,
         instruction_data,
     })
