+++ instruction-macros/crates/instruction-macros/src/derive/instruction_event_data.rs
@@ -1,5 +1,5 @@
-//! Derive helper for generating the `try_from_tag` macro, `pack`, and `unpack` functions for
-//! instruction event data.
+//! Derive helper for generating the `try_from_u8` implementation as well as the pack`, and
+//! `unpack` functions for instruction event data.
 //!
 //! Notably, the structs for these instruction event data types do *not* implement invoke methods,
 //! since they are solely for emitting event data inside a self-CPI instruction.
@@ -12,14 +12,14 @@ use instruction_macros_impl::{
     render::{
         render_instruction_data,
         render_pack_into_slice_trait,
-        render_try_from_tag_macro,
+        render_try_from_u8,
     },
 };
 use proc_macro2::TokenStream;
 use syn::DeriveInput;
 
 pub struct DeriveInstructionEventData {
-    pub try_from_u8_macro: TokenStream,
+    pub try_from_u8: TokenStream,
     pub pack_into_slice_trait: TokenStream,
     pub instruction_data: TokenStream,
 }
@@ -30,12 +30,12 @@ pub fn derive_instruction_event_data(
     let parsed_enum = ParsedEnum::try_from((true, input))?;
     let instruction_variants = parse_instruction_variants(&parsed_enum)?;
 
-    let try_from_u8_macro = render_try_from_tag_macro(&parsed_enum, &instruction_variants);
+    let try_from_u8 = render_try_from_u8(&parsed_enum, &instruction_variants);
     let instruction_data: TokenStream = render_instruction_data(&parsed_enum, instruction_variants);
     let pack_into_slice_trait = render_pack_into_slice_trait();
 
     Ok(DeriveInstructionEventData {
-        try_from_u8_macro,
+        try_from_u8,
         pack_into_slice_trait,
         instruction_data,
     })
